pyproject.toml
[project]
name = "elevation-mapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.4",
    "warp-lang>=1.12.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/elevation_mapping"]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "plotly>=6.7.0",
]
